+++ examples/02_csv_to_alerts.py
@@ -1,3 +1,5 @@
+"""Creates alerts from a csv"""
+
 import csv
 import irflow_client
 import pprint